% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lgb.dl.R
\name{lgb.dl}
\alias{lgb.dl}
\title{Install LightGBM from source}
\usage{
lgb.dl(commit = "master", compiler = "gcc", libdll = "",
  repo = "https://github.com/Microsoft/LightGBM", cores = 1)
}
\arguments{
\item{commit}{The commit / branch to use. Put \code{""} for master branch. Defaults to \code{"master"}.}

\item{compiler}{Applicable only to Windows. The compiler to use (either \code{"gcc"} for MinGW or \code{"vs"} for Visual Studio). Defaults to \code{"gcc"}.}

\item{libdll}{Applicable only if you use a precompiled dll/lib. The precompiled dll/lib to use. Defaults to \code{""}.}

\item{repo}{The link to the repository. Defaults to \code{"https://github.com/Microsoft/LightGBM"}.}

\item{cores}{The number of cores to use for compilation, ignored for Visual Studio. Defaults to \code{1}.}
}
\value{
A logical describing whether the LightGBM package was installed or not (\code{TRUE} if installed, \code{FALSE} if installation failed AND you did not have the package before).
}
\description{
Downloads and install LightGBM from repository. Allows to customize the commit/branch used, the compiler, using a precompiled lib/dll, the link to the repository (if using a custom fork), and the number of cores used (for non-Visual Studio compilations). Requires \code{git} and compiler \code{make} (or \code{mingw32-make} for MinGW) in \code{PATH} environment variable. Windows uses \code{\\\\} (backward slashes) while Linux uses \code{/} (forward slashes).
}
\details{
This installation function supports only Windows (Visual Studio and MinGW) and Linux (gcc and any other compiler making use of \code{make}). Performance of Visual Studio is higher when multithreading is heavy, while MinGW performance is higher when you are using a domestic computer (low number of physical cores, like 2 or 4).

Check here for more details: Laurae's \href{https://github.com/Microsoft/LightGBM/issues/542}{Microsoft/LightGBM#542} (Visual Studio reports higher CPU usage than MinGW). and guolinke's \href{https://github.com/Microsoft/LightGBM/pull/584}{Microsoft/LightGBM#584} (Compile R package by custom tool chain).
}
\examples{
\dontrun{
# Install using Visual Studio
# (Download: http://landinghub.visualstudio.com/visual-cpp-build-tools)
lgb.dl(commit = "master",
       compiler = "vs",
       repo = "https://github.com/Microsoft/LightGBM",
       cores = 4)

# Install using Rtools MinGW or use Linux compilation
lgb.dl(commit = "master",
       compiler = "gcc",
       repo = "https://github.com/Microsoft/LightGBM",
       cores = 4)
       
# Install using precompiled DLL in Windows
lgb.dl(commit = "master",
       libdll = "C:\\\\xgboost\\\\LightGBM\\\\windows\\\\x64\\\\DLL\\\\lib_lightgbm.dll",
       repo = "https://github.com/Microsoft/LightGBM",
       cores = 4)

# Test package
library(lightgbm)
data(agaricus.train, package = "lightgbm")
train <- agaricus.train
dtrain <- lgb.Dataset(train$data, label = train$label)
data(agaricus.test, package = "lightgbm")
test <- agaricus.test
dtest <- lgb.Dataset.create.valid(dtrain, test$data, label = test$label)
params <- list(objective = "regression", metric = "l2")
valids <- list(test = dtest)
model <- lgb.train(params,
                   dtrain,
                   100,
                   valids,
                   min_data = 1,
                   learning_rate = 1,
                   early_stopping_rounds = 10)
}

}
